Move lib/utils/uds to session/uds
