+++ session/networking/networking.go
@@ -46,8 +46,8 @@ import (
 	"github.com/gravitational/trace"
 
 	"github.com/gravitational/teleport/api/defaults"
-	"github.com/gravitational/teleport/lib/utils/uds"
 	"github.com/gravitational/teleport/session/networking/x11"
+	"github.com/gravitational/teleport/session/uds"
 )
 
 // RequestBufferSize is the maximum amount of data we're comfortable writing at
