+++ session/uds/socketpair.go
