+++ session/uds/socketpair_impl_atomic.go
